factor out behaviof the facet context builder for reuse in other resources.  change package to .jsonld
